+++ src/app/home/[username]/layout.tsx
@@ -11,7 +11,7 @@ export const generateMetadata = async ({ params: { username } }: HomeRouteParams
 
 const HomeLayout = ({ children }: PropsWithChildren) => {
   return (
-    <main className="relative flex justify-center bg-gradient1">
+    <main className="relative flex justify-center">
       <div className="absolute right-0 w-full h-[125px] bg-gradientPurpleBlur" />
       <div className="w-full h-[100dvh]">
         <div className="w-full h-[100dvh] flex flex-col items-center justify-between">{children}</div>
