+++ src/app/my/layout.tsx
@@ -1,7 +1,7 @@
 import type { PropsWithChildren } from 'react';
 
 const MyLayout = ({ children }: PropsWithChildren) => {
-  return <div className="w-full h-[100dvh] bg-gradient1">{children}</div>;
+  return <div className="w-full h-[100dvh]">{children}</div>;
 };
 
 export default MyLayout;
